Refactor: decompose api.py and generators to separate files
