+++ dmff/__init__.py
@@ -1,3 +1,4 @@
 from .settings import *
 from .common.nblist import NeighborList, NeighborListFreud
 from .api import Hamiltonian
+from .generators import *
