+++ dmff/generators/__init__.py
@@ -1,0 +1,3 @@
+print("IN API")
+from .admp import *
+from .classical import *
